Revert "Merge branch 'main' into Alta-Cliente"
This reverts commit 6dc53f76b4293f173e2dbc29d8a0178254660828, reversing
changes made to 88fa59053ce7509343f091e29bde26b973feb694.

+++ database/migrations/2023_09_27_170814_create_empleados_table.php
@@ -21,12 +21,12 @@ return new class extends Migration
             $table->integer("dni")->unique();
             $table->string("email");
             $table->string("direccion");
-            $table->bigInteger("telefono");
+            $table->integer("telefono");
             $table->date("fecha_ingreso");
             $table->timestamps();
         });
     }
-    
+
     /**
      * Reverse the migrations.
      */
